Extract parseMarkup into separate module
