+++ src/app/functions/l10n/cronjobs.ts
@@ -5,8 +5,7 @@
 import { dirname, resolve } from "path";
 import { fileURLToPath } from "url";
 import { readFileSync, readdirSync } from "fs";
-import { MarkupParser, ReactLocalization } from "@fluent/react";
-import { JSDOM } from "jsdom";
+import { ReactLocalization } from "@fluent/react";
 import {
   GetL10n,
   createGetL10n,
@@ -15,6 +14,7 @@ import {
   type GetL10nBundles,
 } from ".";
 import type { SanitizedSubscriberRow } from "../server/sanitize";
+import { parseMarkup } from "./parseMarkup";
 
 export function getEmailL10n(
   subscriber: SanitizedSubscriberRow,
@@ -53,26 +53,6 @@ export const getL10nBundles: GetL10nBundles = createGetL10nBundles({
     return ftlPaths.map((filePath) => readFileSync(filePath, "utf-8"));
   },
 });
-
-/**
- * When running in Node.js, Fluent doesn't have access to the DOM APIs to parse HTML in strings. Instead, it calls this function to parse them using JSDOM.
- *
- * See https://github.com/projectfluent/fluent.js/wiki/React-Overlays#custom-markup-parsers
- *
- * @param str Localised string that might include HTML.
- * @returns DOM nodes representing the parsed string.
- */
-const parseMarkup: MarkupParser = (str) => {
-  // Initialising JSDOM for a string that doesn't contain HTML is overkill, so
-  // exit early if the string doesn't even include angle brackets (`<` or `>`):
-  if (!str.includes("<") && !str.includes(">")) {
-    return [{ nodeName: "#text", textContent: str } as Node];
-  }
-  const document = new JSDOM().window.document;
-  const wrapper = document.createElement("span");
-  wrapper.innerHTML = str;
-  return Array.from(wrapper.childNodes);
-};
 
 export const getL10n: GetL10n = createGetL10n({
   getL10nBundles: getL10nBundles,
