+++ src/app/functions/l10n/serverComponents.ts
@@ -4,20 +4,19 @@
 
 import "server-only";
 import { headers } from "next/headers";
-import { JSDOM } from "jsdom";
 import {
   GetL10nBundles,
   createGetL10nBundles,
   GetL10n,
   createGetL10n,
 } from "./index";
-import type { MarkupParser } from "@fluent/react";
 // @fluent/react's default export bundles all code in a single scope, so just
 // importing <ReactLocalization> from there will run createContext,
 // which is not allowed in server components. To avoid that, we import directly
 // from the included ES module code. There is the risk that @fluent/react
 // updates break that.
 import { ReactLocalization } from "@fluent/react/esm/localization";
+import { parseMarkup } from "./parseMarkup";
 
 // `require` isn't usually valid JS, so skip type checking for that:
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -59,27 +58,6 @@ export const getL10nBundles: GetL10nBundles = createGetL10nBundles({
   },
   getAcceptLangHeader: () => headers().get("Accept-Language") ?? "en",
 });
-
-let document: Document;
-/**
- * When running on the server, Fluent doesn't have access to the DOM APIs to parse HTML in strings. Instead, it calls this function to parse them using JSDOM.
- *
- * See https://github.com/projectfluent/fluent.js/wiki/React-Overlays#custom-markup-parsers
- *
- * @param str Localised string that might include HTML.
- * @returns DOM nodes representing the parsed string.
- */
-const parseMarkup: MarkupParser = (str) => {
-  // Initialising JSDOM for a string that doesn't contain HTML is overkill, so
-  // exit early if the string doesn't even include angle brackets (`<` or `>`):
-  if (!str.includes("<") && !str.includes(">")) {
-    return [{ nodeName: "#text", textContent: str } as Node];
-  }
-  document ??= new JSDOM().window.document;
-  const wrapper = document.createElement("span");
-  wrapper.innerHTML = str;
-  return Array.from(wrapper.childNodes);
-};
 
 export const getL10n: GetL10n = createGetL10n({
   getL10nBundles: getL10nBundles,
